requirements.txt
asgiref==3.2.3
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
boto==2.49.0
boto3==1.10.46
botocore==1.13.46
certifi==2019.11.28
cffi==1.13.2
channels==2.4.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
daphne==2.4.1
defusedxml==0.6.0
Django==3.0.1
django-allauth==0.41.0
django-ckeditor==5.9.0
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-js-asset==1.2.2
django-storages==1.8
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docutils==0.15.2
drf-yasg==1.17.0
fcm-django==0.3.4
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
jmespath==0.9.4
jsonfield==2.0.2
MarkupSafe==1.1.1
numpy==1.18.1
oauthlib==3.1.0
packaging==20.1
pandas==0.25.3
Pillow==6.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pyfcm==1.4.7
PyHamcrest==1.9.0
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.7
ruamel.yaml.clib==0.2.0
s3transfer==0.2.1
service-identity==18.1.0
six==1.13.0
sqlparse==0.3.0
txaio==18.8.1
uritemplate==3.0.1
urllib3==1.25.8
xlrd==1.2.0
zope.interface==4.7.1
